Update to Dropzone 5.3.1, comes with browser resizing
Resizing is now done browser-side, although more precise resizing may still be necessary due to scaling and rotation issues (for example, iPhone photos are always landscape with a rotation flag). However, the files sent to the on-device server should reduce bandwidth and computational strain.

+++ phototype/uploader/uploader.js
@@ -9,17 +9,22 @@ var Uploader = {};
 Uploader.init = function () {
 	if (Dropzone && Dropzone.instances[0]) {
 		// adjust some settings on Dropzone instance
-		Dropzone.instances[0].options.addRemoveLinks  = true; // gives option to remove files from queue
-		Dropzone.instances[0].options.acceptedFiles   = 'image/jpeg'; // MIME type
-		Dropzone.instances[0].options.parallelUploads = 1;    // requests active at the same time
-		Dropzone.instances[0].options.maxFilesize     = 20;   // in MB
+		Dropzone.instances[0].options.addRemoveLinks        = true; // gives option to remove files from queue
+		Dropzone.instances[0].options.acceptedFiles         = 'image/jpeg'; // MIME type
+		Dropzone.instances[0].options.parallelUploads       = 1;    // requests active at the same time
+		Dropzone.instances[0].options.maxFilesize           = 20;   // in MB
+		Dropzone.instances[0].options.resizeHeight          = 800;  // resize to given height in pixels
+		Dropzone.instances[0].options.resizeQuality         = 1;    // in range 0..1 (default 0.8)
 
 		// set events on Dropzone instance
 		Dropzone.instances[0].on("success", function (file) {
 			Uploader.onUploadSuccess();
 		});
 		Dropzone.instances[0].on("complete", function (file) {
-			this.removeFile(file);
+			if (file.upload.progress == 100) {
+				this.removeFile(file);
+			}
+			// else something has gone wrong
 		});
 	}
